Update in options_module
we create two functions to calculate the price variation in continues time of an option. The first one receive a pandas series and calculate the return, while the second function receive a pandas series and calculate the cumulative return. The output is type of pandas Series. Both functions were constructed to handle with exceptions like TypeError, ValueError and AttributeError.